# backend/swagger-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Course Final Project - Auction API",
    "description": "Complete API documentation for the online auction platform. This API supports product bidding, user authentication, messaging, and auction management.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "http://localhost:3000"
    },
    "license": {
      "name": "AGPL-3.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "JWT token for authentication"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "description": "Unique user identifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "full_name": {
            "type": "string",
            "description": "Full name of the user"
          },
          "address": {
            "type": "string",
            "description": "User's address"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "description": "Date of birth"
          },
          "role": {
            "type": "string",
            "enum": ["ADMIN", "SELLER", "BIDDER"],
            "description": "User role"
          },
          "is_verified": {
            "type": "boolean",
            "description": "Email verification status"
          },
          "positive_rating": {
            "type": "integer",
            "description": "Count of positive ratings"
          },
          "negative_rating": {
            "type": "integer",
            "description": "Count of negative ratings"
          }
        }
      },
...
      "AutoBid": {
        "type": "object",
        "properties": {
          "auto_bid_id": {
            "type": "integer",
            "description": "Unique auto-bid identifier"
          },
          "product_id": {
            "type": "integer",
            "description": "Product for auto-bidding"
          },
          "bidder_id": {
            "type": "integer",
            "description": "User with auto-bid"
          },
          "max_price": {
            "type": "number",
            "format": "decimal",
            "description": "Maximum price for auto-bidding"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE"],
            "description": "Auto-bid status"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "User Login",
        "operationId": "loginUser",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string",
                      "description": "JWT authentication token"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auto-bids/": {
      "get": {
        "summary": "Get User's Auto Bids",
        "operationId": "getAutoBids",
        "tags": ["Bids"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "query",
            "description": "Filter by product ID (optional)",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Auto bids retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AutoBid"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Auto Bid",
        "operationId": "createAutoBid",
        "tags": ["Bids"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "max_price": {
                    "type": "number",
                    "format": "decimal",
                    "example": 100.00
                  }
                },
                "required": ["product_id", "max_price"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Auto bid created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AutoBid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auto-bids/{id}": {
      "put": {
        "summary": "Update Auto Bid",
        "operationId": "updateAutoBid",
        "tags": ["Bids"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Auto Bid ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "max_price": {
                    "type": "number",
                    "format": "decimal",
                    "example": 150.00
                  }
                },
                "required": ["max_price"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Auto bid updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AutoBid"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}